refactor: import 경로 수정 및 썸네일 이미지 api 값 적용
